+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.77] - 2024-12-19
+
+### Changed
+- Family Tree: Increased person node size to fully display gender icon
+- Family Tree: Added date of birth display in dd-mm-yyyy format on person nodes
+- Family Tree: Improved node spacing and layout
+
 ## [1.0.76] - 2024-12-19
 
 ### Fixed